Avoid detecting a reachable target as unreachable
If the previous target is unreachable and the bot click on a reachable target while the info box "unable to reach..."(from the previous unreachable target) is still on the screen, the bot will think this "unanble to reach..." is for the present reachable target. And therefore the bot will consider the present target, which is reachable, unreachable.

The solution is simple: just sleep for 1 second so the "unable to reach..." info box goes away.