An attempt to solve mis-detection of boss
Bot sometimes cannot detect "unable to reach...". It seems that it occurs at the detection with low similarity 0.8(not sure why;it should be easier). Failure to detect this "unable to reach..." leads to the following situation:

When path to the boss is blocked, the bot cannot see "unable to reach" and blacklist the boss, attacking another target. And then if the bot win, it thinks the boss is defeated and says map clear, stucked at the map.

This might be the reason for the issue "misclassification of the boss" on github.

However, after testing, this mis-detection still occurs.