Fix: Icrease the click attempts of movement for farming 7-2.
If own fleet is far away from the target bot could get stuck because:
If the fleet reaches the target and the number of click attemps just reaches the threshold right before entering the formation, bot will not think it is in the map but is actually in the formation screen, and thereby is stuck.

This can be solved easily by making the threshold very large. But bot will waste a lot of time on clicking.